First push
1. Finished the API request handling and able to print the currently logged in user (logged in via the callback URI) then able to request current user playing track and split the response into SONG NAME by ARTIST NAME

2. Documentation about how to set up a Spotify Developer App included in the python file, and a link back to Spotify documentation about it as well.

TODO: Finish adding actual interaction with OSC so it can be sent to VRChats text box.

Clean up code before release, I just wish to see what's changed since the first push.

Actually have a loop that will continue to see what's playing every x amount of time? or if Spotify has a way to let me know via api the song has change.

Look into if python .env is worth it as a requirement or if just including the secrete and client id in the python file is fine, leaning more towards just leaving it in the file but I don't know yet. A decency just to use .env files seems a bit pointless IMHO.