Uploaded scripts for setting env vars
Uploaded 2 scripts for setting environment variables on Windows, one written in batch and the other written in powershell

Included a persistent variables shell script for both macOS and Linux and one that is non persistent that will only apply the variables to the current shell if that is for whatever reason the desired effect.

Included a readme for github / if the user opens and reads it on their machines, if that isn't the case a tl;dr esque txt file is also included.